requirements.txt
aerich==0.7.1
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dictdiffer==0.9.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
magic-filter==1.0.9
multidict==6.0.4
pydantic==1.10.9
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
tomlkit==0.11.8
tortoise-orm==0.19.3
typing_extensions==4.6.3
yarl==1.9.2
